Now shows program images in program list. Add seperate script for generating test users. Some changes in broadcast model
