+++ server/package.json
@@ -5,7 +5,9 @@
   "scripts": {
     "start": "babel-node app/index.js",
     "test": "mocha **/*.spec.js --compilers js:babel-register",
-    "load-example-data": "babel-node generate_example_data.js"
+    "load-example-data": "babel-node generate_example_data.js",
+    "import-from-chimera": "babel-node import_from_chimera.js",
+    "generate-test-users": "babel-node generate_test_users.js"
   },
   "author": "",
   "private": true,
